requirements.txt
einops==0.4.1
hydra-core==1.2.0
imageio==2.9.0
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.2.2
opencv_python==4.6.0.66
Pillow==9.2.0
pytorch3d==0.6.2
pyzmq==23.2.0
rawpy==0.17.1
requests==2.27.1
scikit_image==0.19.2
scipy==1.7.3
skimage==0.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
visdom==0.1.8.9
